add recurring events functionality

+++ app/controllers/events_controller.rb
@@ -14,7 +14,7 @@ class EventsController < ApplicationController
       hash = {user_name: user.name, color: user.events_color.present? ? user.events_color : user.color(index)}
       events = []
       user.events.each do |event|
-        event = {id: event.id, title: event.title, start: event.starts_at, end: event.ends_at, allDay: event.all_day}
+        event = {id: event.id, title: event.title, start: event.starts_at, end: event.ends_at, allDay: event.all_day }
         events << event
       end
       hash[:events] = events
@@ -110,12 +110,17 @@ class EventsController < ApplicationController
                         nylas_namespace_id: n_event.namespace_id, namespace_id: calendar.namespace_id,
                         when_type: n_event.when['object'])
         #TODO figure out how to send recuring events to nylas
-        # elsif @event.class.name == 'EventSeries'
-        #   n_event = nylas_namespace.events.build(:calendar_id => calendar.calendar_id, :title => @event.title, :description => @event.description,
-        #                                          :location => @event.location, :when => {:start_time => @event.starts_at.to_i,
-        #                                                                                  :end_time => @event.ends_at.to_i},
-        #                                          :participants => @event.events.first.participants.map {|p| { :email => p.email, :name => p.name}})
-        #   n_event.save!
+        elsif @event.class.name == 'EventSeries'
+          @event.events.each do |e|
+            n_event = nylas_namespace.events.build(:calendar_id => calendar.calendar_id, :title => e.title, :description => e.description,
+                                                   :location => e.location, :when => {:start_time => e.starts_at.to_i,
+                                                                                           :end_time => e.ends_at.to_i},
+                                                   :participants => e.participants.map {|p| { :email => p.email, :name => p.name}})
+            n_event.save!
+            e.update(calendar_id: calendar.id, nylas_event_id: n_event.id, nylas_calendar_id: n_event.calendar_id,
+                          nylas_namespace_id: n_event.namespace_id, namespace_id: calendar.namespace_id,
+                          when_type: n_event.when['object'])
+          end
         end
       end
     end
@@ -137,28 +142,36 @@ class EventsController < ApplicationController
         ends_at: event_params[:all_day] ? Date.strptime(event_params[:end_date], '%m/%d/%Y').strftime('%Y-%m-%d') : DateTime.strptime("#{event_params[:end_date]} #{event_params[:end_time]}", '%m/%d/%Y %H:%M %p').strftime('%Y-%m-%d %H:%M %p'),
         all_day: event_params[:all_day]
     }
-    if @event.update(attrs)
-      participants = event_params[:participants].split(",") if event_params[:participants].present?
-      if participants.present?
-        participants.each do |p_email|
-          participant = Participant.find_or_create_by(email: p_email)
-          event_participant = EventParticipant.create(event_id: @event.id, participant_id: participant.id)
+    if @event.event_series.present? && event_params[:update_all_events] == 'true'
+      @event.event_series.update_events_until_end_time(event_params)
+    else
+      if @event.update(attrs)
+        participants = event_params[:participants].split(",") if event_params[:participants].present?
+        if participants.present?
+          event_participants = @event.event_participants.where(event_id: @event.id).to_a
+          participants.each do |p_email|
+            participant = Participant.find_or_create_by(email: p_email)
+            event_participant = EventParticipant.find_or_create_by(event_id: @event.id, participant_id: participant.id)
+            event_participants.delete(event_participant)
+          end
+          event_participants.each {|ep| ep.destroy }
+        end
+        if calendar.present?
+          namespace = calendar.namespace
+          @inbox = Nylas::API.new(Rails.application.secrets.inbox_app_id, Rails.application.secrets.inbox_app_secret, namespace.inbox_token)
+          nylas_namespace = @inbox.namespaces.first
+          n_event = nylas_namespace.events.find(@event.nylas_event_id)
+          n_event.title = @event.title
+          n_event.description = @event.description
+          n_event.location = @event.location
+          n_event.when = {:start_time => @event.starts_at.to_i, :end_time => @event.ends_at.to_i}
+          n_event.participants = @event.participants.map {|p| { :email => p.email, :name => p.name}}
+          n_event.save!
+          @event.update(when_type: n_event.when['object']) if n_event.when['object'] != @event.when_type
         end
       end
-      if calendar.present?
-        namespace = calendar.namespace
-        @inbox = Nylas::API.new(Rails.application.secrets.inbox_app_id, Rails.application.secrets.inbox_app_secret, namespace.inbox_token)
-        nylas_namespace = @inbox.namespaces.first
-        n_event = nylas_namespace.events.find(@event.nylas_event_id)
-        n_event.title = @event.title
-        n_event.description = @event.description
-        n_event.location = @event.location
-        n_event.when = {:start_time => @event.starts_at.to_i, :end_time => @event.ends_at.to_i}
-        n_event.participants = @event.participants.map {|p| { :email => p.email, :name => p.name}}
-        n_event.save!
-        @event.update(when_type: n_event.when['object']) if n_event.when['object'] != @event.when_type
-      end
     end
+
     message = @event.errors.present? ? {error: @event.errors.full_messages.to_sentence} : {notice: 'Event was successfully updated.'}
     respond_to do |format|
       format.html {redirect_to request.referrer , :flash => message}
@@ -188,14 +201,29 @@ class EventsController < ApplicationController
   # DELETE /events/1
   # DELETE /events/1.json
   def destroy
-    calendar = @event.calendar
-    nylas_event_id = @event.nylas_event_id
-    if @event.destroy
-      if nylas_event_id.present? && calendar.present?
-        namespace = calendar.namespace
-        @inbox = Nylas::API.new(Rails.application.secrets.inbox_app_id, Rails.application.secrets.inbox_app_secret, namespace.inbox_token)
-        nylas_namespace = @inbox.namespaces.first
-        n_event = nylas_namespace.events.find(nylas_event_id).destroy
+    if params[:delete_all] == 'true' && @event.event_series.present?
+      @event.event_series.events.each do |e|
+        calendar = e.calendar
+        nylas_event_id = e.nylas_event_id
+        if e.destroy
+          if nylas_event_id.present? && calendar.present?
+            namespace = calendar.namespace
+            @inbox = Nylas::API.new(Rails.application.secrets.inbox_app_id, Rails.application.secrets.inbox_app_secret, namespace.inbox_token)
+            nylas_namespace = @inbox.namespaces.first
+            n_event = nylas_namespace.events.find(nylas_event_id).destroy
+          end
+        end
+      end
+    else
+      calendar = @event.calendar
+      nylas_event_id = @event.nylas_event_id
+      if @event.destroy
+        if nylas_event_id.present? && calendar.present?
+          namespace = calendar.namespace
+          @inbox = Nylas::API.new(Rails.application.secrets.inbox_app_id, Rails.application.secrets.inbox_app_secret, namespace.inbox_token)
+          nylas_namespace = @inbox.namespaces.first
+          n_event = nylas_namespace.events.find(nylas_event_id).destroy
+        end
       end
     end
     respond_to do |format|
@@ -261,9 +289,6 @@ class EventsController < ApplicationController
                     end
                   end
                 end
-              elsif event == "delete"
-                event = Event.find_by(user_id: @user.id, nylas_event_id: ne.id).try(:destroy)
-                #TODO add async event removing from calendar
               end
               events_synced += 1
             last_cursor = ne.cursor
@@ -277,7 +302,7 @@ class EventsController < ApplicationController
       hash = {user_name: user.name, color: user.events_color.present? ? user.events_color : user.color(index)}
       events = []
       user.events.each do |event|
-        event = {id: event.id, title: event.title, start: event.starts_at, end: event.ends_at, allDay: event.when_type == 'date' || event.when_type == 'datespan'}
+        event = {id: event.id, title: event.title, start: event.starts_at, end: event.ends_at, allDay: event.all_day}
         events << event
       end
       hash[:events] = events
@@ -315,7 +340,7 @@ class EventsController < ApplicationController
   def event_params
     params.require(:event).permit(:title, :location, :description, :calendar_id, :summary, :start_date, :start_time,
                                   :end_date, :end_time, :all_day, :status, :participants, :recur, :period, :frequency,
-                                  :recur_start_date, :recur_end_date)
+                                  :recur_start_date, :recur_end_date, :event_series_id, :update_all_events)
   end
 
   def event_drag_params
